CORENET-6324: Add automation for ovnk branch sync (#70537)
* re-org OWNERS file w/ symlinks

adding a second folder here for branch-sync work so just need
one OWNERS file with symlinks from each subfolder

Signed-off-by: Jamo Luhrsen <jluhrsen@gmail.com>

* add branch sync automation for ovnk

adds branch sync automation via 'git merge -x theirs' for:

4.20->4.19
4.19->4.18

Signed-off-by: Jamo Luhrsen <jluhrsen@gmail.com>

* fix argtype in d/s merge script for draft prs

while debugging the new branch-sync sript it was noticed that
the same code borrowed from the downstream merge script to
post a draft PR was not working right. The fix was to use
argjson for the draft argument. this fixes that

Signed-off-by: Jamo Luhrsen <jluhrsen@gmail.com>

---------

Signed-off-by: Jamo Luhrsen <jluhrsen@gmail.com>