steps.osc: Improve ARO workflow (#72042)
* steps.osc: Remove different-region-pp-handling

this handling comes from upstream script, which is using comunity image
published only in eastus. Downstream we either select the right region
(for CoCo) or the operator builds it's own image in the expected region,
so we don't need this handling. Let's remove it to always test the
supported configuration.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>

* steps.osc: Add support for ARO

this workflow is similar to azure but uses the MS ARO flavour. As
deleting of the ARO cluster might fail when podvms are still running
this patch adds a best-effort cleanup.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>

* steps.osc: Allow to override ARO region

To override ARO location we need to modify the HYPERSHIFT_AZURE_LOCATION
and then add support to set the LOCATION in aro* steps, which are
already respecting this variables just not specifying it in their
metadata.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>

* steps.osc: Fix INITDATA default

prow linter requires "" for empty string and not ''.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>

---------

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>