step-registry/sandboxed-containers-operator: strengthening get-kata-rpm step (#72067)
* sandboxed-containers-operator/get-kata-rpm: add more prints and debugs

Some times the step fails and no information is printed in the logs.
Let's add some prints before important actions so that at least we have
a better idea where it failed.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>

* sandboxed-containers-operator/get-kata-rpm: set shell errexit/nounset/pipefail

Run the script in a more strict mode: any error should stop the
execution. Initialize the exported environment so nounset does not fail.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>

* sandboxed-containers-operator/get-kata-rpm: save curl exit code properly

Use `||` to capture the return code of curl command once the script runs on errexit mode.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>

* sandboxed-containers-operator/get-kata-rpm: hard-code base URLs

Have the base url hard-coded in the step, instead of reading from
secrets. It helps with debugging and testing the step locally.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>

* sandboxed-containers-operator/get-kata-rpm: fix curl'ed file checking

If curl doesn't fail then it will look for messages in
kata-containers.rpm that indicates the file was not found. Fix that
checking because it's always passing.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>

* sandboxed-containers-operator/get-kata-rpm: new BREW_AUTH variable

If BREW_AUTH is exported then it will use that instead of getting from
openshift-config.

BREW_AUTH should not be a parameter of the step. It's only used for
debugging and testing the step locally.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>

* sandboxed-containers-operator/get-kata-rpm: check workers are found

If workers are not found then there is a bug and the script should bail
out.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>

* sandboxed-containers-operator/get-kata-rpm: refactor md5sum checking

Avoid grep'ing the md5sum string, just compare the strings.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>

---------

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>